+++ tests/test_cli.py
@@ -52,15 +52,15 @@ def test_console_entry_point() -> None:
         (
             ["", "--mqtt-host", "mqtt-broker.local"],
             "mqtt-broker.local",
-            1883,
+            8883,
             None,
             None,
             3,
         ),
         (
-            ["", "--mqtt-host", "mqtt-broker.local", "--mqtt-port", "8883"],
+            ["", "--mqtt-host", "mqtt-broker.local", "--mqtt-port", "8884"],
             "mqtt-broker.local",
-            8883,
+            8884,
             None,
             None,
             3,
@@ -68,7 +68,7 @@ def test_console_entry_point() -> None:
         (
             ["", "--mqtt-host", "mqtt-broker.local", "--mqtt-username", "me"],
             "mqtt-broker.local",
-            1883,
+            8883,
             "me",
             None,
             3,
@@ -86,7 +86,7 @@ def test_console_entry_point() -> None:
                 "21",
             ],
             "mqtt-broker.local",
-            1883,
+            8883,
             "me",
             "secret",
             21,
@@ -103,12 +103,12 @@ def test__main(
 ) -> None:
     with unittest.mock.patch("switchbot_mqtt._run") as run_mock, unittest.mock.patch(
         "sys.argv", argv
-    ), pytest.warns(UserWarning, match=r"Please add --mqtt-disable-tls\b"):
+    ):
         switchbot_mqtt._cli._main()
     run_mock.assert_called_once_with(
         mqtt_host=expected_mqtt_host,
         mqtt_port=expected_mqtt_port,
-        mqtt_disable_tls=True,
+        mqtt_disable_tls=False,
         mqtt_username=expected_username,
         mqtt_password=expected_password,
         mqtt_topic_prefix="homeassistant/",
@@ -152,8 +152,8 @@ def test__main_mqtt_password_file(
         switchbot_mqtt._cli._main()
     run_mock.assert_called_once_with(
         mqtt_host="localhost",
-        mqtt_port=1883,
-        mqtt_disable_tls=True,
+        mqtt_port=8883,
+        mqtt_disable_tls=False,
         mqtt_username="me",
         mqtt_password=expected_password,
         mqtt_topic_prefix="homeassistant/",
@@ -215,8 +215,8 @@ def test__main_device_password_file(
         switchbot_mqtt._cli._main()
     run_mock.assert_called_once_with(
         mqtt_host="localhost",
-        mqtt_port=1883,
-        mqtt_disable_tls=True,
+        mqtt_port=8883,
+        mqtt_disable_tls=False,
         mqtt_username=None,
         mqtt_password=None,
         mqtt_topic_prefix="homeassistant/",
@@ -228,8 +228,8 @@ def test__main_device_password_file(
 
 _RUN_DEFAULT_KWARGS: typing.Dict[str, typing.Any] = {
     "mqtt_host": "localhost",
-    "mqtt_port": 1883,
-    "mqtt_disable_tls": True,
+    "mqtt_port": 8883,
+    "mqtt_disable_tls": False,
     "mqtt_username": None,
     "mqtt_password": None,
     "mqtt_topic_prefix": "homeassistant/",
@@ -239,10 +239,10 @@ _RUN_DEFAULT_KWARGS: typing.Dict[str, typing.Any] = {
 }
 
 
-def test__main_mqtt_disable_tls_implicit() -> None:
+def test__main_mqtt_disable_tls() -> None:
     with unittest.mock.patch("switchbot_mqtt._run") as run_mock, unittest.mock.patch(
-        "sys.argv", ["", "--mqtt-host", "mqtt.local"]
-    ), pytest.warns(UserWarning, match=r"Please add --mqtt-disable-tls\b"):
+        "sys.argv", ["", "--mqtt-host", "mqtt.local", "--mqtt-disable-tls"]
+    ):
         switchbot_mqtt._cli._main()
     run_mock.assert_called_once_with(
         **{
@@ -254,47 +254,19 @@ def test__main_mqtt_disable_tls_implicit() -> None:
     )
 
 
-def test__main_mqtt_enable_tls() -> None:
+def test__main_mqtt_disable_tls_overwrite_port() -> None:
     with unittest.mock.patch("switchbot_mqtt._run") as run_mock, unittest.mock.patch(
-        "sys.argv", ["", "--mqtt-host", "mqtt.local", "--mqtt-enable-tls"]
+        "sys.argv",
+        ["", "--mqtt-host", "mqtt.local", "--mqtt-port", "1884", "--mqtt-disable-tls"],
     ):
         switchbot_mqtt._cli._main()
     run_mock.assert_called_once_with(
         **{
             **_RUN_DEFAULT_KWARGS,
             "mqtt_host": "mqtt.local",
-            "mqtt_disable_tls": False,
-            "mqtt_port": 8883,
+            "mqtt_disable_tls": True,
+            "mqtt_port": 1884,
         }
-    )
-
-
-def test__main_mqtt_enable_tls_overwrite_port() -> None:
-    with unittest.mock.patch("switchbot_mqtt._run") as run_mock, unittest.mock.patch(
-        "sys.argv",
-        ["", "--mqtt-host", "mqtt.local", "--mqtt-port", "1883", "--mqtt-enable-tls"],
-    ):
-        switchbot_mqtt._cli._main()
-    run_mock.assert_called_once_with(
-        **{
-            **_RUN_DEFAULT_KWARGS,
-            "mqtt_host": "mqtt.local",
-            "mqtt_disable_tls": False,
-            "mqtt_port": 1883,
-        }
-    )
-
-
-def test__main_mqtt_tls_collision(capsys: _pytest.capture.CaptureFixture) -> None:
-    with unittest.mock.patch("switchbot_mqtt._run") as run_mock, unittest.mock.patch(
-        "sys.argv",
-        ["", "--mqtt-host", "mqtt.local", "--mqtt-enable-tls", "--mqtt-disable-tls"],
-    ), pytest.raises(SystemExit):
-        switchbot_mqtt._cli._main()
-    run_mock.assert_not_called()
-    assert (
-        "error: argument --mqtt-disable-tls: not allowed with argument --mqtt-enable-tls\n"
-        in capsys.readouterr()[1]
     )